+++ packages/form-footer/src/utils/getLgIds.ts
@@ -7,8 +7,11 @@ export const getLgIds = (root: LgIdString = DEFAULT_LGID_ROOT) => {
     root,
     backButton: `${root}-back_button`,
     backButtonIcon: `${root}-back_button-icon`,
+    backSplitButton: `${root}-back_split_button`,
     cancelButton: `${root}-cancel_button`,
+    cancelSplitButton: `${root}-cancel_split_button`,
     primaryButton: `${root}-primary_button`,
+    primarySplitButton: `${root}-primary_split_button`,
   } as const;
   return ids;
 };