+++ chat/chat-disclaimer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @lg-chat/chat-disclaimer
 
+## 4.0.4
+
+### Patch Changes
+
+- @leafygreen-ui/marketing-modal@6.0.4
+
+## 4.0.3
+
+### Patch Changes
+
+- Updated dependencies [b67497a]
+  - @leafygreen-ui/lib@15.2.0
+  - @leafygreen-ui/marketing-modal@6.0.3
+  - @leafygreen-ui/tokens@3.1.2
+  - @leafygreen-ui/typography@22.0.1
+
 ## 4.0.2
 
 ### Patch Changes