+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @leafygreen-ui/modal
 
+## 19.0.1
+
+### Patch Changes
+
+- Updated dependencies [b67497a]
+  - @leafygreen-ui/lib@15.2.0
+  - @leafygreen-ui/hooks@9.1.1
+  - @leafygreen-ui/icon@14.1.0
+  - @leafygreen-ui/icon-button@17.0.3
+  - @leafygreen-ui/leafygreen-provider@5.0.2
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/portal@7.0.2
+  - @leafygreen-ui/tokens@3.1.2
+
 ## 19.0.0
 
 ### Major Changes