+++ packages/number-input/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @leafygreen-ui/number-input
 
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies [2adba55]
+  - @leafygreen-ui/button@25.0.0
+  - @leafygreen-ui/form-field@4.0.1
+  - @leafygreen-ui/select@16.0.2
+  - @leafygreen-ui/tooltip@14.1.1
+
+## 5.0.3
+
+### Patch Changes
+
+- Updated dependencies [b67497a]
+  - @leafygreen-ui/lib@15.2.0
+  - @leafygreen-ui/a11y@3.0.2
+  - @leafygreen-ui/button@24.0.3
+  - @leafygreen-ui/form-field@4.0.1
+  - @leafygreen-ui/hooks@9.1.1
+  - @leafygreen-ui/icon@14.1.0
+  - @leafygreen-ui/leafygreen-provider@5.0.2
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/select@16.0.1
+  - @leafygreen-ui/tokens@3.1.2
+  - @leafygreen-ui/tooltip@14.1.1
+  - @leafygreen-ui/typography@22.0.1
+
 ## 5.0.2
 
 ### Patch Changes