+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leafygreen-ui/popover
 
+## 14.0.3
+
+### Patch Changes
+
+- Updated dependencies [b67497a]
+  - @leafygreen-ui/lib@15.2.0
+  - @leafygreen-ui/hooks@9.1.1
+  - @leafygreen-ui/leafygreen-provider@5.0.2
+  - @leafygreen-ui/portal@7.0.2
+  - @leafygreen-ui/tokens@3.1.2
+
 ## 14.0.2
 
 ### Patch Changes