+++ packages/radio-group/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/radio-group
 
+## 13.0.3
+
+### Patch Changes
+
+- Updated dependencies [b67497a]
+  - @leafygreen-ui/lib@15.2.0
+  - @leafygreen-ui/hooks@9.1.1
+  - @leafygreen-ui/leafygreen-provider@5.0.2
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/tokens@3.1.2
+  - @leafygreen-ui/typography@22.0.1
+
 ## 13.0.2
 
 ### Patch Changes