+++ packages/split-button/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @leafygreen-ui/split-button
 
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies [2adba55]
+  - @leafygreen-ui/button@25.0.0
+  - @leafygreen-ui/menu@31.0.1
+  - @leafygreen-ui/popover@14.0.3
+
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies [b67497a]
+  - @leafygreen-ui/lib@15.2.0
+  - @leafygreen-ui/button@24.0.3
+  - @leafygreen-ui/hooks@9.1.1
+  - @leafygreen-ui/icon@14.1.0
+  - @leafygreen-ui/leafygreen-provider@5.0.2
+  - @leafygreen-ui/menu@31.0.1
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/polymorphic@3.0.3
+  - @leafygreen-ui/popover@14.0.3
+  - @leafygreen-ui/tokens@3.1.2
+
 ## 6.0.0
 
 ### Major Changes