+++ packages/typography/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/typography
 
+## 22.0.1
+
+### Patch Changes
+
+- Updated dependencies [b67497a]
+  - @leafygreen-ui/lib@15.2.0
+  - @leafygreen-ui/icon@14.1.0
+  - @leafygreen-ui/leafygreen-provider@5.0.2
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/polymorphic@3.0.3
+  - @leafygreen-ui/tokens@3.1.2
+
 ## 22.0.0
 
 ### Major Changes